requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
SQLAlchemy==2.0.23
python-wordpress-xmlrpc==2.3
beautifulsoup4==4.12.2
html2text==2020.1.16
pytz==2023.3
google-generativeai==0.3.0
markdownify==0.11.6
markdown==3.5.1
pygments==2.17.2